Remove unneeded effect modifiers
The `endMatchInTurn`  closure already produces an
effect. Which to my understanding means there is no need to
`catchToEffect` anymore, making that a useless noop.
Furthermore, if we `fireAndForget`, then ignoring any
output and errors is already implied and again, the call
to `ignoreOutput` is just noise. As with the above
`eraseToEffect` was a noop from the beginning
since everything is already an effect.